move httpserver under server
